move brands definition
